backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
supabase==2.19.0
httpx==0.27.0
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
